fix(FontParser): signed ToF2Dot14, BigEndianReader bounds checks, ReadNullTerminatedString
- ToF2Dot14(ushort): F2Dot14 is signed 2.14 fixed-point; reinterpret the bits as a
  signed short before scaling (0xC000 = -1.0, was +3.0). Affects mirrored composite
  transforms.
- BigEndianReader.ReadByte/PeekBytes/ReadSbytes: add bounds checks that throw past the
  logical end instead of silently returning ArrayPool slack (stale garbage). These are
  used by live table parsers, so a truncated/malformed table now fails loudly (and is
  caught by PdfLibrary's per-font guards) rather than parsing garbage.
- ReadNullTerminatedString(unicode): the loop compared array references
  (`PeekBytes(2) != new byte[]{0,0}`, always true) and decoded UTF-16 little-endian.
  Fixed to terminate on a 0x0000 unit and decode big-endian (sfnt convention).

All 25 ClaimProbes green (reader + post + sfnt + cff); 293 PDF tests pass; GlyphDiff
unchanged at 6/26,630 diverged (no regression from the bounds checks).

+++ FontParser/Extensions/UshortExtensions.cs
@@ -9,7 +9,9 @@
 
         public static float ToF2Dot14(this ushort value)
         {
-            return value / 16384f;
+            // F2Dot14 is a SIGNED 2.14 fixed-point value: reinterpret the 16 bits as a
+            // signed short before scaling (e.g. 0xC000 = -1.0, not +3.0).
+            return (short)value / 16384f;
         }
     }
 }
